+++ gluecodium/src/main/java/com/here/gluecodium/generator/swift/SwiftNameRules.kt
@@ -22,6 +22,7 @@ package com.here.gluecodium.generator.swift
 import com.here.gluecodium.generator.common.NameRuleSet
 import com.here.gluecodium.generator.common.NameRules
 import com.here.gluecodium.model.lime.LimeAttributeType.SWIFT
+import com.here.gluecodium.model.lime.LimeAttributeValueType
 import com.here.gluecodium.model.lime.LimeAttributeValueType.NAME
 import com.here.gluecodium.model.lime.LimeElement
 import com.here.gluecodium.model.lime.LimeNamedElement
@@ -35,12 +36,16 @@ class SwiftNameRules(nameRuleSet: NameRuleSet) : NameRules(nameRuleSet) {
     override fun getPropertyName(limeProperty: LimeProperty) =
         getPlatformName(limeProperty) ?: super.getPropertyName(limeProperty)
 
-    fun getImplementationFileName(limeElement: LimeNamedElement) =
-        (TARGET_DIRECTORY +
+    fun getImplementationFileName(limeElement: LimeNamedElement): String {
+        val isSwiftExtension = limeElement.attributes.have(SWIFT, LimeAttributeValueType.EXTENSION)
+        val suffix = if (isSwiftExtension) "__extension" else ""
+        return (TARGET_DIRECTORY +
                 limeElement.path.head.joinToString(File.separator) +
                 File.separator +
                 getName(limeElement) +
+                suffix +
                 ".swift")
+    }
 
     private fun getPlatformName(limeElement: LimeNamedElement?) =
         limeElement?.attributes?.get(SWIFT, NAME)