+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Features:
   + Added validation against broken import statements in LimeIDL.
+  + Added support for `@Swift(Extension)` attribute for type collections (`types`) in LimeIDL.
 
 ## 5.7.0
 - Release date: 2019-10-25